--- src/main/java/faang/school/postservice/exception/comment/DataValidationException.java
@@ -1,7 +1,0 @@
-package faang.school.postservice.exception.comment;
-
-public class DataValidationException extends RuntimeException {
-    public DataValidationException(String message) {
-        super(message);
-    }
-}
