+++ src/main/java/faang/school/postservice/mapper/CommentMapper.java
@@ -1,6 +1,6 @@
-package faang.school.postservice.mapper.comment;
+package faang.school.postservice.mapper;
 
-import faang.school.postservice.dto.comment.CommentDto;
+import faang.school.postservice.dto.CommentDto;
 import faang.school.postservice.model.Comment;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
