+++ src/test/java/faang/school/postservice/mapper/CommentMapperTest.java
@@ -1,7 +1,8 @@
-package faang.school.postservice.mapper.comment;
+package faang.school.postservice.mapper;
 
-import faang.school.postservice.dto.comment.CommentDto;
+import faang.school.postservice.dto.CommentDto;
 import faang.school.postservice.dto.UserDto;
+import faang.school.postservice.mapper.CommentMapper;
 import faang.school.postservice.model.Comment;
 import faang.school.postservice.model.Like;
 import faang.school.postservice.model.Post;
