+++ src/test/java/faang/school/postservice/service/CommentServiceTest.java
@@ -1,15 +1,15 @@
-package faang.school.postservice.service.comment;
+package faang.school.postservice.service;
 
 import faang.school.postservice.client.UserServiceClient;
+import faang.school.postservice.dto.CommentDto;
 import faang.school.postservice.dto.UserDto;
-import faang.school.postservice.dto.comment.CommentDto;
-import faang.school.postservice.exception.comment.DataValidationException;
-import faang.school.postservice.mapper.comment.CommentMapper;
+import faang.school.postservice.exception.DataValidationException;
+import faang.school.postservice.mapper.CommentMapper;
 import faang.school.postservice.model.Comment;
 import faang.school.postservice.model.Post;
 import faang.school.postservice.repository.CommentRepository;
 import faang.school.postservice.repository.PostRepository;
-import faang.school.postservice.validation.comment.CommentValidator;
+import faang.school.postservice.validator.CommentValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
